Rename print-without-types to strip-types
The module no longer exports printWithoutTypes(). Rename file, test,
imports, and stale comments to match the new API surface.
