+++ packages/jsx/src/strip-types.ts
